requirements.txt
pymongo
psycopg2-binary
pandas
faker
apache-airflow[mongo,postgres]
streamlit
plotly
sqlalchemy